backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
llama-parse
llama-index
pinecone-client>=3.0.0
sentence-transformers
pydantic
python-dotenv
requests
crawl4ai
playwright
aiofiles
cors
fastapi-cors
langchain-openai>=0.1.0
langchain-community>=0.2.0
google-search-results>=2.4.2
beautifulsoup4==4.12.2
# Authentication dependencies
PyJWT>=2.8.0
email-validator>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
# Supabase
supabase>=2.0.0
postgresql-adapter>=0.1.0
psycopg2-binary>=2.9.0